removed interface prom provenance msg provider
